refactor: decompose opdrachten-sidebar into focused sub-components
Extract the 1614-line opdrachten-sidebar.tsx into 11 focused modules
under components/sidebar/:
- sidebar-types.ts: shared types, interfaces, and style constants
- sidebar-utils.ts: utility functions (search, parse, URL helpers)
- use-sidebar-filters.ts: custom hook for all filter state and query logic
- sidebar-search-bar.tsx: search input (compact + overview variants)
- sidebar-filter-controls.tsx: FilterChecklist, CompactMultiSelectFilter, RadiusSliderField
- sidebar-sort-controls.tsx: sort dropdown
- sidebar-results-header.tsx: results count, badges, page size selector
- sidebar-pagination.tsx: pagination controls
- sidebar-job-list.tsx: job list with ScrollArea
- compact-sidebar-filters.tsx: compact dark-themed filter grid
- overview-filter-panel.tsx: full overview filter panel

Container is now 268 lines (was 1614). No behavioral changes.
